+++ rssparser/src/main/java/com/niilopoutanen/rssparser/NewParser.java
@@ -7,8 +7,15 @@ import com.niilopoutanen.rssparser.parsers.RssParser;
 
 import org.jsoup.nodes.Document;
 
+import java.time.ZonedDateTime;
+import java.time.format.DateTimeFormatter;
+import java.time.format.DateTimeParseException;
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
+import java.util.Locale;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 public class NewParser {
     public Source source;
@@ -17,9 +24,9 @@ public class NewParser {
     public NewParser(){
 
     }
-
     public void get(String url){
         Document document = WebUtils.connect(url);
+        parse(document);
     }
 
     public void parse(Document document){
@@ -34,6 +41,35 @@ public class NewParser {
             atomParser.parse(document);
             source = atomParser.getSource();
             posts = atomParser.getPosts();
+        }
+    }
+
+    public static Date parseDate(String dateString){
+        List<DateTimeFormatter> formats = new ArrayList<>();
+        formats.add(DateTimeFormatter.ofPattern("E, d MMM yyyy HH:mm:ss Z", Locale.ENGLISH));
+        formats.add(DateTimeFormatter.ofPattern("E, d MMM yyyy HH:mm:ss zzz", Locale.ENGLISH));
+        formats.add(DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ssXXX"));
+        formats.add(DateTimeFormatter.ofPattern("EEE, dd MMM yyyy HH:mm:ss Z", Locale.ENGLISH));
+
+        for(DateTimeFormatter formatter : formats){
+            try{
+                ZonedDateTime zonedDateTime = ZonedDateTime.parse(dateString, formatter);
+                return Date.from(zonedDateTime.toInstant());
+            }
+            catch (DateTimeParseException ignored){}
+        }
+
+        return null;
+    }
+    public static String parsePattern(String raw, String attribute){
+        String regexPattern = attribute + "=\"(.*?)\"";
+        Pattern pattern = Pattern.compile(regexPattern);
+        Matcher matcher = pattern.matcher(raw);
+
+        if (matcher.find()) {
+            return matcher.group(1);
+        } else {
+            return "";
         }
     }
 }
